Revert "Re-organize utils"
