+++ test/test_construct/test_contains_domain.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tgrag.construct_graph_scripts.construct_subnetwork import (
+from tgrag.construct_graph_scripts.subnetwork_construct import (
     contains_base_domain,
 )
