+++ test/test_construct/test_subnetwork_construction.py
@@ -1,9 +1,12 @@
 import pandas as pd
 import pytest
 
-from tgrag.construct_graph_scripts.construct_subnetwork import (
+from tgrag.construct_graph_scripts.subnetwork_construct import (
     construct_subnetwork,
 )
+
+# TO DO: this file does not have type annotations 
+
 
 @pytest.fixture
 def test_data(tmp_path):
