+++ tgrag/construct_graph_scripts/load_labels.py
@@ -4,30 +4,12 @@ from pathlib import Path
 import pandas as pd
 
 from tgrag.utils.matching import extract_graph_domains
-from tgrag.utils.path import get_root_dir
 
 
 def load_credibility_scores(path: str, use_core: bool = False) -> pd.DataFrame:
     cred_df = pd.read_csv(path)
     cred_df['match_domain'] = cred_df['domain']
     return cred_df[['match_domain', 'pc1']]
-
-
-def get_labelled_set() -> set[str]:
-    """Get a list (set) of labelled domains."""
-    path = os.path.join(get_root_dir(), 'data', 'dqr', 'domain_pc1.csv')
-    wanted_domains = set()
-
-    with open(path, 'r', encoding='utf-8') as f:
-        for line in f:
-            parts = line.strip().split(',')
-            if len(parts) >= 1:
-                domain = parts[0].strip()
-                if domain:  # skip empty lines
-                    wanted_domains.add(domain)
-
-    # print(f'[INFO] Found {len(wanted_domains)} domains ')
-    return wanted_domains
 
 
 def get_credibility_intersection(
